Extract About page into focused reusable subcomponents
